🔊 Feat Debug log would not output when debugmode is disable.
